// gen/ts/open_splunk/history_api.ts
export enum SearchHistorySortBy {
  SEARCH_HISTORY_SORT_BY_UNSPECIFIED = 0,
  SEARCH_HISTORY_SORT_BY_CREATED_AT = 1,
  SEARCH_HISTORY_SORT_BY_FINISHED_AT = 2,
  SEARCH_HISTORY_SORT_BY_DURATION = 3,
  SEARCH_HISTORY_SORT_BY_MATCHED_EVENTS = 4,
  UNRECOGNIZED = -1,
}